requirements.txt
apache-airflow==2.9.2
apache-airflow-providers-common-sql==1.3.4
apache-airflow-providers-postgres==5.11.1
apache-airflow-providers-http==4.1.0
psycopg2-binary==2.9.7
pandas==1.5.3
SQLAlchemy==1.4.52
python-dotenv==1.0.0
requests==2.28.1
pyarrow==10.0.1



  


